[PB_1766]:(feature) Reorganize webdav server files
