+++ package.json
@@ -10,7 +10,7 @@
     "prepack": "yarn build && oclif manifest && oclif readme",
     "prepare": "husky",
     "test:unit": "nyc --reporter=lcov --reporter=text mocha \"test/**/*.test.ts\"",
-    "dev:webdav": "nodemon -e ts --exec 'ts-node src/webdav/server.ts'",
+    "dev:webdav": "nodemon -e ts --exec 'ts-node src/webdav/index.ts'",
     "version": "oclif readme && git add README.md"
   },
   "homepage": "https://github.com/internxt/cli",
